[AS-606] fix imports
