+++ src/index.ts
@@ -1,6 +1,6 @@
 export * from "./services";
 export * from "./domain";
-export * from "./config";
-export * from "./utils";
+export * from "./infrastructure/configuration";
+export * from "./infrastructure/misc";
 export * from "./infrastructure/adapters";
 export * from "./application";
