+++ src/infrastructure/adapters/Rholang/DeployFactory/shared.ts
@@ -1,0 +1,8 @@
+export const escapeRholangString = (value: string): string => {
+    return value
+        .replace(/\\/g, "\\\\")
+        .replace(/"/g, '\\"')
+        .replace(/\n/g, "\\n")
+        .replace(/\r/g, "\\r")
+        .replace(/\t/g, "\\t");
+};
